Add files via upload
This is a sentiment analysis program that uses a recursive neural tensor network (RNTN) to take as input sentences from movie reviews and then to score them (1-5) on how positive or negative the sentences are. The data files are all from the Stanford Natural Language Processing (NLP) group.

More specifically, in my implementation, each sentence is processed into a tree first, then turned into a sequence with post-order traversal, then fed into a recurrent neural network (RNN) for predicting and training. The reason for this iterative approach as opposed to the more intuitive recursive approach is that Theano builds a graph of all the shared variables before doing computations with them, so a recursive approach would incur heavy memory cost.